// mini-xrm-frontend/mini-xrm-client-angular/model/createOrUpdatePartnerRequestView.ts
/**
 * Mini xRM API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PartnerStatusView } from './partnerStatusView';


export interface CreateOrUpdatePartnerRequestView { 
    name: string;
    taxNumber: string;
    headquarters?: string;
    status: PartnerStatusView;
    tagIds: Array<number>;
}
export namespace CreateOrUpdatePartnerRequestView {
}